fix: データモデルやフォーマッター、ユーティリティなどの切り出し
