+++ app/Sources/Features/HeatMapFeature.swift
@@ -41,41 +41,3 @@ struct HeatMapFeature: Reducer {
         }
     }
 }
-
-struct FocusData: Equatable, Identifiable, Decodable {
-    let id = UUID()
-    let date: Date
-    let hour: Int
-    let focus_score: Int
-}
-
-struct FocusDataWrapper: Decodable {
-    let items: [FocusData]
-}
-
-enum DummyDataLoader {
-    static func loadFocusData() -> [FocusData] {
-        guard let url = Bundle.main.url(forResource: "heatmap_data", withExtension: "json"),
-              let data = try? Data(contentsOf: url) else {
-            return []
-        }
-
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .formatted(DateFormatter.yyyyMMdd)
-
-        if let decoded = try? decoder.decode(FocusDataWrapper.self, from: data) {
-            return decoded.items
-        }
-
-        return []
-    }
-}
-
-private extension DateFormatter {
-    static let yyyyMMdd: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        formatter.timeZone = .current
-        return formatter
-    }()
-}
